+++ truffle/src/com.oracle.truffle.api.test/src/com/oracle/truffle/api/test/interop/ForeignAccessToStringTest.java
@@ -48,8 +48,6 @@ import com.oracle.truffle.api.CallTarget;
 import com.oracle.truffle.api.interop.ForeignAccess;
 import com.oracle.truffle.api.interop.Message;
 import com.oracle.truffle.api.interop.TruffleObject;
-import com.oracle.truffle.api.nodes.RootNode;
-import java.util.function.Supplier;
 
 public class ForeignAccessToStringTest {
     @Test
@@ -60,7 +58,7 @@ public class ForeignAccessToStringTest {
 
     @Test
     public void check10Factory() {
-        ForeignAccess fa = ForeignAccess.create(new Simple10TestingFactory(), (Supplier<RootNode>) null);
+        ForeignAccess fa = ForeignAccess.createAccess(new Simple10TestingFactory(), null);
         assertEquals("ForeignAccess[" + ForeignAccessToStringTest.class.getName() + "$Simple10TestingFactory]", fa.toString());
     }