+++ truffle/src/com.oracle.truffle.dsl.processor.interop/src/com/oracle/truffle/dsl/processor/interop/ForeignAccessFactoryGenerator.java
@@ -151,6 +151,7 @@ final class ForeignAccessFactoryGenerator {
         imports.add("com.oracle.truffle.api.CallTarget");
         if (hasLanguageCheckNode()) {
             imports.add("com.oracle.truffle.api.CompilerDirectives.TruffleBoundary");
+            imports.add("java.util.function.Supplier");
         }
         imports.add("com.oracle.truffle.api.Truffle");
         imports.add("com.oracle.truffle.api.interop.ForeignAccess");
@@ -158,7 +159,6 @@ final class ForeignAccessFactoryGenerator {
         imports.add("com.oracle.truffle.api.interop.ForeignAccess.StandardFactory");
         imports.add("com.oracle.truffle.api.interop.Message");
         imports.add("com.oracle.truffle.api.interop.TruffleObject");
-        imports.add("java.util.function.Supplier");
         if (!(messageGenerators.containsKey(Message.IS_BOXED) &&
                         messageGenerators.containsKey(Message.IS_NULL) &&
                         messageGenerators.containsKey(Message.IS_EXECUTABLE) &&
@@ -188,11 +188,11 @@ final class ForeignAccessFactoryGenerator {
     }
 
     private void appendSingletonAndGetter(Writer w) throws IOException {
-        String allocation = "ForeignAccess.create(new " + simpleClassName + "(), ";
+        String allocation = "ForeignAccess.createAccess(new " + simpleClassName + "(), ";
         if (hasLanguageCheckNode()) {
             allocation += "new Supplier<RootNode>() { @Override public RootNode get() { return " + languageCheckGenerator.getRootNodeFactoryInvocation() + "; }});";
         } else {
-            allocation += "(Supplier<RootNode>) null);";
+            allocation += "null);";
         }
         w.append("    public static final ForeignAccess ACCESS = ").append(allocation).append("\n");
         w.append("    @Deprecated public static ForeignAccess createAccess() { return ").append(allocation).append(" }\n");